Add RFC 7830 padding to DNS requests (#51)
Add `Message.AddPadding()` which automatically adds padding to the DNS
message. The padding size is calculated to make the total size of each
message a multiple of 128 as recommended in RFC 8467.

This change also adds partial support for RFC 6891 - Extension
Mechanisms for DNS (EDNS(0)). In particular, when padding is used,
servers may return extended response codes. This change also adds
`Message.ResponseCode()` , which returns the rcode or extended rcode
automatically.