Allow comma-separated list of addresses in Dial (#53)
The `addr` argument to `Dial` is now a comma-separated list of
addresses.

The existing behavior hasn't changed. `Dial` resolves the first address
and attempts to connect to each resolved target until it connects
successfully. If `Dial` runs out of targets to try, it moves on to the
next address, etc.